pyproject.toml
[project]
name = "repl-nix-new-repl"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.0",
    "openai>=1.57.4",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
]
